// src/model/user.ts
export interface User {

    /**
     * the user id generated by the database
     * @readonly
     */
    readonly id: number;

    /**
     * the creation datetime
     * @readonly
     */
    readonly created: Date;

    /**
     * the modified datetime
     * @readonly
     */
    readonly modified: Date;

    /**
     * the user name (nickname)
     */
    username: string;

    /**
     * the user email address
     */
    email: string;

}